// src/types/report.d.ts
export interface IncidentLog {
  id: string;
  incidentId: string;
  hazardId: string | null;
  status: "ASSIGNED" | "COMPLETED" | "PENDING" | "IN_PROGRESS" | string; // extend as needed
  userId: string;
  comment: string;
  createdAt: string; // ISO date string
  updatedAt: string; // ISO date string
}
export interface Comment {
  id: string;
  comment?: string;
  followUpDescription?: string;
  createdAt: string; // ISO date string
  name: string;
  email: string;
  role: string;
  providedImageUrl: string;
  phoneNumber: string;
  userId: string;
}

export interface ReportResponse {
  report: Report;
  incident?: Incident;
  hazard?: Hazard;
  incidentAssignee: AssigneeUser;
  dynamicQuestion: {
    question: FormQuestion;
    answer: FormAnswer;
  }[];
  media: IncidentMedia[];
  logs: IncidentLog[];
  reportLogs: IncidentLog[];
  comments: Comment[];
  followUps?: Comment[];
}

// Report object
export interface Report {
  id: string;
  title: string;
  description: string;
  mainType: "INCIDENT" | "HAZARD";
  status: "INITIATED" | "IN_PROGRESS" | "RESOLVED" | string;
  priority: "LOW" | "MEDIUM" | "HIGH" | "EXTREME" | string;
  hazardId: string | null;
  incidentId: string;
  incident_coordinates: string;
  userId: string;
  updatedBy: string;
  createdAt: string;
  updatedAt: string;
  followUp: boolean;
}

// Incident object
export interface Incident {
  id: string;
  name: string;
  incidentDescription: string;
  status: "INITIATED" | "IN_PROGRESS" | "RESOLVED" | string;
  treatmentType: "FIRST_AID" | "NO_TREATMENT_PROVIDED" | "OTHER" | string;
  treatmentDescription: string;
  injuredBodyPart: string;
  groupId: string | null;
  createdAt: string;
  updatedAt: string;
}
export interface Hazard {
  id: string;
  name: string;
  hazardDescription: string;
  status: "INITIATED" | "IN_PROGRESS" | "RESOLVED" | string;
  groupId: string | null;
  createdAt: string;
  updatedAt: string;
}

// Media (attached to incident status log)
export interface IncidentMedia {
  id: string;
  mediaId: string;
  incidentStatusLogId: string;
  createdAt: string;
  updatedAt: string;
  status: "INITIATED" | "IN_PROGRESS" | "RESOLVED" | string;
  comment: string;
  url: string;
}

// Incident Assignee
export interface AssigneeUser {
  id: string;
  name: string;
  email: string;
  role: "DEPARTMENT_MANAGER" | "STAFF"; // Add more roles if needed
  providedImageUrl: string;
  phoneNumber: string;
  assigntype: "SELF_ASSIGNED" | "MANAGER_ASSIGNED";
}
// ...
export enum treatmentType {
  FIRST_AID = "FIRST_AID",
  NO_TREATMENT_PROVIDED = "NO_TREATMENT_PROVIDED",
  MONITORED = "MONITORED",
  DOCTOR_GP = "DOCTOR_GP",
  HOSPITAL = "HOSPITAL",
  OTHER = "OTHER",
}